+++ mlxtend/preprocessing/dense_transformer.py
@@ -19,6 +19,20 @@ class DenseTransformer(BaseEstimator):
         self.is_fitted = False
 
     def transform(self, X, y=None):
+        """ Return a dense version of the input array.
+
+        Parameters
+        ----------
+        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
+            Training vectors, where n_samples is the number of samples and
+            n_features is the number of features.
+        y : array-like, shape = [n_samples] (default: None)
+
+        Returns
+        ---------
+        X_dense : dense version of the input X array.
+
+        """
         if issparse(X):
             return X.toarray()
         elif self.return_copy:
@@ -27,8 +41,36 @@ class DenseTransformer(BaseEstimator):
             return X
 
     def fit(self, X, y=None):
+        """ Mock method. Does nothing.
+
+        Parameters
+        ----------
+        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
+            Training vectors, where n_samples is the number of samples and
+            n_features is the number of features.
+        y : array-like, shape = [n_samples] (default: None)
+
+        Returns
+        ---------
+        self
+
+        """
         self.is_fitted = True
         return self
 
     def fit_transform(self, X, y=None):
+        """ Return a dense version of the input array.
+
+        Parameters
+        ----------
+        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
+            Training vectors, where n_samples is the number of samples and
+            n_features is the number of features.
+        y : array-like, shape = [n_samples] (default: None)
+
+        Returns
+        ---------
+        X_dense : dense version of the input X array.
+
+        """
         return self.transform(X=X, y=y)