+++ docs/sources/CHANGELOG.md
@@ -11,7 +11,8 @@
 
 ##### New Features
 
-~
+- Added `preprocessing.CopyTransformer`, a mock class that returns copies of
+imput arrays via `transform` and `fit_transform`
 
 ##### Changes
 
@@ -23,6 +24,7 @@
 - Fixed rare bug in MLP layer_mapping functions that caused a swap between the random number generation seed when initializing weights and biases
 - More rigorous type and shape checks in `evaluate.plot_decision_regions`
 - Changes in `DenseTransformer` so that it doesn't fail if the input array is not sparse
+- Use scikit-learn's `BaseEstimator` as parent class for `feature_selection.ColumnSelector`
 
 ### Version 0.4.1 (2016-05-01)